requirements.txt
--find-links https://download.pytorch.org/whl/cu113/torch_stable.html
torch>=1.10.0+cu113
torchvision>=0.9.0+cu113
pytorch-lightning>=1.2.3
pandas>=1.2.5
numpy>=1.20.1
pillow>=8.1.2
requests
